+++ spec/cached_json_spec.rb
@@ -2,8 +2,8 @@ require 'spec_helper'
 
 describe Mongoid::CachedJson do
   it 'has a version' do
-    Mongoid::CachedJson::VERSION.should_not be_nil
-    Mongoid::CachedJson::VERSION.to_f.should > 0
+    expect(Mongoid::CachedJson::VERSION).not_to be_nil
+    expect(Mongoid::CachedJson::VERSION.to_f).to be > 0
   end
 
   [:dalli_store, :memory_store].each do |cache_store|
@@ -23,14 +23,14 @@ describe Mongoid::CachedJson do
       context 'with basic fields defined for export with json_fields' do
         it 'returns public JSON if you nil options' do
           example = JsonFoobar.create(foo: 'FOO', baz: 'BAZ', bar: 'BAR')
-          example.as_json(nil).should == example.as_json(properties: :short)
+          expect(example.as_json(nil)).to eq(example.as_json(properties: :short))
         end
         it 'allows subsets of fields to be returned by varying the properties definition' do
           example = JsonFoobar.create(foo: 'FOO', baz: 'BAZ', bar: 'BAR')
           # :short is a subset of the fields in :public and :public is a subset of the fields in :all
-          example.as_json(properties: :short).should == { :foo => 'FOO', 'Baz' => 'BAZ', :default_foo => 'DEFAULT_FOO' }
-          example.as_json(properties: :public).should == { :foo => 'FOO', 'Baz' => 'BAZ', :bar => 'BAR', :default_foo => 'DEFAULT_FOO' }
-          example.as_json(properties: :all).should == { :foo => 'FOO', :bar => 'BAR', 'Baz' => 'BAZ', :renamed_baz => 'BAZ', :default_foo => 'DEFAULT_FOO', :computed_field => 'FOOBAR' }
+          expect(example.as_json(properties: :short)).to eq(:foo => 'FOO', 'Baz' => 'BAZ', :default_foo => 'DEFAULT_FOO')
+          expect(example.as_json(properties: :public)).to eq(:foo => 'FOO', 'Baz' => 'BAZ', :bar => 'BAR', :default_foo => 'DEFAULT_FOO')
+          expect(example.as_json(properties: :all)).to eq(:foo => 'FOO', :bar => 'BAR', 'Baz' => 'BAZ', :renamed_baz => 'BAZ', :default_foo => 'DEFAULT_FOO', :computed_field => 'FOOBAR')
         end
         it 'throws an error if you ask for an undefined property type' do
           expect { JsonFoobar.create.as_json(properties: :special) }.to raise_error(ArgumentError)
@@ -43,12 +43,12 @@ describe Mongoid::CachedJson do
           all_result = foobar.as_json(properties: :all)
           public_result = foobar.as_json(properties: :public)
           short_result = foobar.as_json(properties: :short)
-          all_result.should_not == public_result
-          all_result.should_not == short_result
-          public_result.should_not == short_result
-          3.times { foobar.as_json(properties: :all).should == all_result }
-          3.times { foobar.as_json(properties: :public).should == public_result }
-          3.times { foobar.as_json(properties: :short).should == short_result }
+          expect(all_result).not_to eq(public_result)
+          expect(all_result).not_to eq(short_result)
+          expect(public_result).not_to eq(short_result)
+          3.times { expect(foobar.as_json(properties: :all)).to eq(all_result) }
+          3.times { expect(foobar.as_json(properties: :public)).to eq(public_result) }
+          3.times { expect(foobar.as_json(properties: :short)).to eq(short_result) }
         end
         it 'should remove values from the cache when a model is saved' do
           foobar = JsonFoobar.create(foo: 'FOO', bar: 'BAR', baz: 'BAZ')
@@ -57,13 +57,13 @@ describe Mongoid::CachedJson do
           short_result = foobar.as_json(properties: :short)
           foobar.foo = 'updated'
           # Not saved yet, so we should still be hitting the cache
-          3.times { foobar.as_json(properties: :all).should == all_result }
-          3.times { foobar.as_json(properties: :public).should == public_result }
-          3.times { foobar.as_json(properties: :short).should == short_result }
+          3.times { expect(foobar.as_json(properties: :all)).to eq(all_result) }
+          3.times { expect(foobar.as_json(properties: :public)).to eq(public_result) }
+          3.times { expect(foobar.as_json(properties: :short)).to eq(short_result) }
           foobar.save
-          3.times { foobar.as_json(properties: :all).should == all_result.merge(foo: 'updated', computed_field: 'updatedBAR') }
-          3.times { foobar.as_json(properties: :public).should == public_result.merge(foo: 'updated') }
-          3.times { foobar.as_json(properties: :short).should == short_result.merge(foo: 'updated') }
+          3.times { expect(foobar.as_json(properties: :all)).to eq(all_result.merge(foo: 'updated', computed_field: 'updatedBAR')) }
+          3.times { expect(foobar.as_json(properties: :public)).to eq(public_result.merge(foo: 'updated')) }
+          3.times { expect(foobar.as_json(properties: :short)).to eq(short_result.merge(foo: 'updated')) }
         end
       end
       context 'invalidate callbacks' do
@@ -95,36 +95,36 @@ describe Mongoid::CachedJson do
           3.times do
             3.times do
               manager_short_json = manager.as_json(properties: :short)
-              manager_short_json.length.should == 2
-              manager_short_json[:name].should == 'Boss'
-              manager_short_json[:employees].member?(name: 'Peon').should be_true
-              manager_short_json[:employees].member?(name: 'Indentured servant').should be_true
-              manager_short_json[:employees].member?(name: 'Serf').should be_true
-              manager_short_json[:employees].member?(nickname: 'Serf').should be_false
+              expect(manager_short_json.length).to eq(2)
+              expect(manager_short_json[:name]).to eq('Boss')
+              expect(manager_short_json[:employees].member?(name: 'Peon')).to be_truthy
+              expect(manager_short_json[:employees].member?(name: 'Indentured servant')).to be_truthy
+              expect(manager_short_json[:employees].member?(name: 'Serf')).to be_truthy
+              expect(manager_short_json[:employees].member?(nickname: 'Serf')).to be_falsey
             end
             3.times do
               manager_public_json = manager.as_json(properties: :public)
-              manager_public_json.length.should == 2
-              manager_public_json[:name].should == 'Boss'
-              manager_public_json[:employees].member?(name: 'Peon').should be_true
-              manager_public_json[:employees].member?(name: 'Indentured servant').should be_true
-              manager_public_json[:employees].member?(name: 'Serf').should be_true
-              manager_public_json[:employees].member?(nickname: 'Serf').should be_false
+              expect(manager_public_json.length).to eq(2)
+              expect(manager_public_json[:name]).to eq('Boss')
+              expect(manager_public_json[:employees].member?(name: 'Peon')).to be_truthy
+              expect(manager_public_json[:employees].member?(name: 'Indentured servant')).to be_truthy
+              expect(manager_public_json[:employees].member?(name: 'Serf')).to be_truthy
+              expect(manager_public_json[:employees].member?(nickname: 'Serf')).to be_falsey
             end
             3.times do
               manager_all_json = manager.as_json(properties: :all)
-              manager_all_json.length.should == 3
-              manager_all_json[:name].should == 'Boss'
-              manager_all_json[:ssn].should == '123-45-6789'
-              manager_all_json[:employees].member?(name: 'Peon', nickname: 'My Favorite').should be_true
-              manager_all_json[:employees].member?(name: 'Indentured servant', nickname: 'My Favorite').should be_true
-              manager_all_json[:employees].member?(name: 'Serf', nickname: 'Vince').should be_true
+              expect(manager_all_json.length).to eq(3)
+              expect(manager_all_json[:name]).to eq('Boss')
+              expect(manager_all_json[:ssn]).to eq('123-45-6789')
+              expect(manager_all_json[:employees].member?(name: 'Peon', nickname: 'My Favorite')).to be_truthy
+              expect(manager_all_json[:employees].member?(name: 'Indentured servant', nickname: 'My Favorite')).to be_truthy
+              expect(manager_all_json[:employees].member?(name: 'Serf', nickname: 'Vince')).to be_truthy
             end
             3.times do
-              peon.as_json(properties: :short).should == { name: 'Peon' }
+              expect(peon.as_json(properties: :short)).to eq(name: 'Peon')
             end
             3.times do
-              peon.as_json(properties: :all).should == { name: 'Peon', nickname: 'My Favorite' }
+              expect(peon.as_json(properties: :all)).to eq(name: 'Peon', nickname: 'My Favorite')
             end
           end
         end
@@ -132,22 +132,22 @@ describe Mongoid::CachedJson do
           manager = JsonManager.create(name: 'JsonManager')
           employee = manager.json_employees.create(name: 'JsonEmployee')
           3.times do
-            manager.as_json(properties: :short).should == { name: 'JsonManager', employees: [{ name: 'JsonEmployee' }] }
-            employee.as_json(properties: :short).should == { name: 'JsonEmployee' }
+            expect(manager.as_json(properties: :short)).to eq(name: 'JsonManager', employees: [{ name: 'JsonEmployee' }])
+            expect(employee.as_json(properties: :short)).to eq(name: 'JsonEmployee')
           end
           manager.name = 'New JsonManager'
           manager.save
-          3.times { manager.as_json(properties: :short).should == { name: 'New JsonManager', employees: [{ name: 'JsonEmployee' }] } }
+          3.times { expect(manager.as_json(properties: :short)).to eq(name: 'New JsonManager', employees: [{ name: 'JsonEmployee' }]) }
           employee.name = 'New JsonEmployee'
           employee.save
-          3.times { manager.as_json(properties: :short).should == { name: 'New JsonManager', employees: [{ name: 'New JsonEmployee' }] } }
+          3.times { expect(manager.as_json(properties: :short)).to eq(name: 'New JsonManager', employees: [{ name: 'New JsonEmployee' }]) }
         end
         context 'reference_properties' do
           it 'limits the json fields of a child relationship' do
             supervisor = JsonSupervisor.create(name: 'JsonSupervisor')
             manager = JsonManager.create(name: 'JsonManager', supervisor: supervisor)
             json = supervisor.as_json(properties: :all)
-            json[:managers][0].key?(:ssn).should be_false
+            expect(json[:managers][0].key?(:ssn)).to be_falsey
           end
         end
       end
@@ -157,7 +157,7 @@ describe Mongoid::CachedJson do
         end
         it 'uses the correct properties on the base object and passes :all to any sub-objects for :all properties' do
           3.times do
-            @artwork.as_json(properties: :all).should == { name: 'Mona Lisa', image: nil }
+            expect(@artwork.as_json(properties: :all)).to eq(name: 'Mona Lisa', image: nil)
           end
         end
         context 'with the relationship present' do
@@ -166,16 +166,16 @@ describe Mongoid::CachedJson do
           end
           it 'uses the correct properties on the base object and passes :short to any sub-objects for :public and :short properties' do
             3.times do
-              @artwork.as_json(properties: :short).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona' } }
-              @artwork.as_json(properties: :public).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona' } }
-              @artwork.as_json(properties: :all).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona', url: 'http://example.com/404.html' } }
-              @image.as_json(properties: :short).should == { name: 'Picture of Mona Lisa', nickname: 'Mona' }
-              @image.as_json(properties: :public).should == { name: 'Picture of Mona Lisa', nickname: 'Mona', url: 'http://example.com/404.html' }
+              expect(@artwork.as_json(properties: :short)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona' })
+              expect(@artwork.as_json(properties: :public)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona' })
+              expect(@artwork.as_json(properties: :all)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona', url: 'http://example.com/404.html' })
+              expect(@image.as_json(properties: :short)).to eq(name: 'Picture of Mona Lisa', nickname: 'Mona')
+              expect(@image.as_json(properties: :public)).to eq(name: 'Picture of Mona Lisa', nickname: 'Mona', url: 'http://example.com/404.html')
             end
           end
           it 'uses the correct properties on the base object and passes :all to any sub-objects for :all properties' do
             3.times do
-              @artwork.as_json(properties: :all).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona', url: 'http://example.com/404.html' } }
+              expect(@artwork.as_json(properties: :all)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona', url: 'http://example.com/404.html' })
             end
           end
           it 'correctly updates fields when either the parent or child class changes' do
@@ -184,27 +184,27 @@ describe Mongoid::CachedJson do
             @image.nickname = 'Worst Painting Ever'
             # Nothing has been saved yet, cached json for referenced document should reflect the truth in the database
             3.times do
-              @artwork.as_json(properties: :short).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona' } }
-              @artwork.as_json(properties: :public).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona' } }
-              @artwork.as_json(properties: :all).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona', url: 'http://example.com/404.html' } }
+              expect(@artwork.as_json(properties: :short)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona' })
+              expect(@artwork.as_json(properties: :public)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona' })
+              expect(@artwork.as_json(properties: :all)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Mona', url: 'http://example.com/404.html' })
             end
             @image.save
             3.times do
-              @artwork.as_json(properties: :short).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever' } }
-              @artwork.as_json(properties: :public).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever' } }
-              @artwork.as_json(properties: :all).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever', url: 'http://example.com/404.html' } }
+              expect(@artwork.as_json(properties: :short)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever' })
+              expect(@artwork.as_json(properties: :public)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever' })
+              expect(@artwork.as_json(properties: :all)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever', url: 'http://example.com/404.html' })
             end
             @image.name = 'Picture of Mona Lisa Watercolor'
             3.times do
-              @artwork.as_json(properties: :short).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever' } }
-              @artwork.as_json(properties: :public).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever' } }
-              @artwork.as_json(properties: :all).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever', url: 'http://example.com/404.html' } }
+              expect(@artwork.as_json(properties: :short)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever' })
+              expect(@artwork.as_json(properties: :public)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever' })
+              expect(@artwork.as_json(properties: :all)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa', nickname: 'Worst Painting Ever', url: 'http://example.com/404.html' })
             end
             @image.save
             3.times do
-              @artwork.as_json(properties: :short).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa Watercolor', nickname: 'Worst Painting Ever' } }
-              @artwork.as_json(properties: :public).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa Watercolor', nickname: 'Worst Painting Ever' } }
-              @artwork.as_json(properties: :all).should == { name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa Watercolor', nickname: 'Worst Painting Ever', url: 'http://example.com/404.html' } }
+              expect(@artwork.as_json(properties: :short)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa Watercolor', nickname: 'Worst Painting Ever' })
+              expect(@artwork.as_json(properties: :public)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa Watercolor', nickname: 'Worst Painting Ever' })
+              expect(@artwork.as_json(properties: :all)).to eq(name: 'Mona Lisa', image: { name: 'Picture of Mona Lisa Watercolor', nickname: 'Worst Painting Ever', url: 'http://example.com/404.html' })
             end
           end
         end
@@ -212,26 +212,26 @@ describe Mongoid::CachedJson do
       context 'with a hide_as_child_json_when definition' do
         it 'should yield JSON when as_json is called directly and hide_as_child_json_when returns false on an instance' do
           c = SometimesSecret.create(should_tell_secret: true)
-          c.as_json(properties: :short).should == { secret: 'Afraid of the dark' }
+          expect(c.as_json(properties: :short)).to eq(secret: 'Afraid of the dark')
         end
         it 'should yield JSON when as_json is called directly and hide_as_child_json_when returns true on an instance' do
           c = SometimesSecret.create(should_tell_secret: false)
-          c.as_json(properties: :short).should == { secret: 'Afraid of the dark' }
+          expect(c.as_json(properties: :short)).to eq(secret: 'Afraid of the dark')
         end
         it 'should yield JSON without an instance of a child' do
           p = SecretParent.create(name: 'Parent')
-          p.as_json(properties: :all)[:child].should be_nil
+          expect(p.as_json(properties: :all)[:child]).to be_nil
         end
         it 'should yield child JSON when as_json is called on the parent and hide_as_child_json_when returns false on an instance' do
           p = SecretParent.create(name: 'Parent')
           p.create_sometimes_secret(should_tell_secret: true)
-          p.as_json(properties: :short)[:child].should == { secret: 'Afraid of the dark' }
+          expect(p.as_json(properties: :short)[:child]).to eq(secret: 'Afraid of the dark')
         end
         it 'should not yield child JSON when as_json is called on the parent and hide_as_child_json_when returns true on an instance' do
           p = SecretParent.create(name: 'Parent')
           p.create_sometimes_secret(should_tell_secret: false)
-          p.as_json(properties: :short).should == { name: 'Parent', child: nil }
-          p.as_json(properties: :short)[:child].should be_nil
+          expect(p.as_json(properties: :short)).to eq(name: 'Parent', child: nil)
+          expect(p.as_json(properties: :short)[:child]).to be_nil
         end
       end
       context 'relationships with a multi-level hierarchy' do
@@ -245,7 +245,7 @@ describe Mongoid::CachedJson do
         end
         it 'uses the correct properties on the base object and passes :short to any sub-objects for :short and :public' do
           3.times do
-            @artwork.as_json(properties: :short).should == {
+            expect(@artwork.as_json(properties: :short)).to eq(
               name: 'Artwork',
               image: { name: 'Image',
                        urls: [
@@ -254,8 +254,8 @@ describe Mongoid::CachedJson do
                          { url: @common_url }
                        ]
               }
-            }
-            @artwork.as_json(properties: :public).should == {
+            )
+            expect(@artwork.as_json(properties: :public)).to eq(
               name: 'Artwork',
               display_name: 'Awesome Artwork',
               image: { name: 'Image',
@@ -265,12 +265,12 @@ describe Mongoid::CachedJson do
                          { url: @common_url }
                        ]
               }
-            }
+            )
           end
         end
         it 'uses the correct properties on the base object and passes :all to any sub-objects for :all' do
           3.times do
-            @artwork.as_json(properties: :all).should == {
+            expect(@artwork.as_json(properties: :all)).to eq(
               name: 'Artwork',
               display_name: 'Awesome Artwork',
               price: 1000,
@@ -280,7 +280,7 @@ describe Mongoid::CachedJson do
                          { url: @common_url, is_public: false },
                          { url: @common_url, is_public: false }]
                 }
-            }
+            )
           end
         end
         it 'correctly updates json for all classes in the hierarchy when saves occur' do
@@ -290,7 +290,7 @@ describe Mongoid::CachedJson do
           @url1.url = new_url
           # No save has happened, so as_json shouldn't update yet
           3.times do
-            @artwork.as_json(properties: :short).should == {
+            expect(@artwork.as_json(properties: :short)).to eq(
               name: 'Artwork',
               image: { name: 'Image',
                        urls: [
@@ -299,14 +299,14 @@ describe Mongoid::CachedJson do
                          { url: @common_url }
                        ]
               }
-            }
+            )
           end
           @url1.save
           3.times do
             json = @artwork.as_json
-            json[:name].should == 'Artwork'
-            json[:image][:name].should == 'Image'
-            json[:image][:urls].map { |u| u[:url] }.sort.should == [@common_url, @common_url, new_url].sort
+            expect(json[:name]).to eq('Artwork')
+            expect(json[:image][:name]).to eq('Image')
+            expect(json[:image][:urls].map { |u| u[:url] }.sort).to eq([@common_url, @common_url, new_url].sort)
           end
         end
       end
@@ -318,7 +318,7 @@ describe Mongoid::CachedJson do
             end
           end
           it 'transforms every value in returned JSON' do
-            JsonFoobar.new(foo: 'foo', bar: 'Bar', baz: 'BAZ').as_json.should == { 'Baz' => 'BAZ', :default_foo => 'DEFAULT_FOO', :foo => 'FOO' }
+            expect(JsonFoobar.new(foo: 'foo', bar: 'Bar', baz: 'BAZ').as_json).to eq('Baz' => 'BAZ', :default_foo => 'DEFAULT_FOO', :foo => 'FOO')
           end
         end
         context 'with options' do
@@ -328,7 +328,7 @@ describe Mongoid::CachedJson do
             end
           end
           it 'transforms every value in returned JSON using the :transform attribute' do
-            JsonTransform.new(upcase: 'upcase', downcase: 'DOWNCASE', nochange: 'eLiTe').as_json.should == { upcase: 'UPCASE', downcase: 'downcase', nochange: 'eLiTe' }
+            expect(JsonTransform.new(upcase: 'upcase', downcase: 'DOWNCASE', nochange: 'eLiTe').as_json).to eq(upcase: 'UPCASE', downcase: 'downcase', nochange: 'eLiTe')
           end
         end
         context 'with mutliple transformations' do
@@ -341,22 +341,22 @@ describe Mongoid::CachedJson do
             end
           end
           it 'transforms every value in returned JSON using the :transform attribute' do
-            JsonMath.new(number: 9).as_json.should == { number: 5 }
+            expect(JsonMath.new(number: 9).as_json).to eq(number: 5)
           end
         end
       end
       context 'with cache disabled' do
         before :each do
-          Mongoid::CachedJson.config.stub(:disable_caching).and_return(true)
+          allow(Mongoid::CachedJson.config).to receive(:disable_caching).and_return(true)
         end
         it 'forces a cache miss' do
           example = JsonFoobar.create(foo: 'FOO', baz: 'BAZ', bar: 'BAR')
           key = "as_json/unspecified/JsonFoobar/#{example.id}/short/true"
           case cache_store
           when :memory_store then
-            Mongoid::CachedJson.config.cache.should_receive(:fetch).with(key,  force: true).twice
+            expect(Mongoid::CachedJson.config.cache).to receive(:fetch).with(key,  force: true).twice
           when :dalli_store then
-            Mongoid::CachedJson.config.cache.should_not_receive(:write)
+            expect(Mongoid::CachedJson.config.cache).not_to receive(:write)
           else
             fail ArgumentError, "invalid cache store: #{cache_store}"
           end
@@ -366,26 +366,26 @@ describe Mongoid::CachedJson do
       context 'versioning' do
         it 'returns JSON for version 2' do
           example = JsonFoobar.create(foo: 'FOO', baz: 'BAZ', bar: 'BAR')
-          example.as_json(properties: :short, version: :v2).should == { :foo => 'FOO', 'Taz' => 'BAZ', 'Naz' => 'BAZ', :default_foo => 'DEFAULT_FOO' }
+          expect(example.as_json(properties: :short, version: :v2)).to eq(:foo => 'FOO', 'Taz' => 'BAZ', 'Naz' => 'BAZ', :default_foo => 'DEFAULT_FOO')
         end
         it 'returns JSON for version 3' do
           example = JsonFoobar.create(foo: 'FOO', baz: 'BAZ', bar: 'BAR')
-          example.as_json(properties: :short, version: :v3).should == { :foo => 'FOO', 'Naz' => 'BAZ', :default_foo => 'DEFAULT_FOO' }
+          expect(example.as_json(properties: :short, version: :v3)).to eq(:foo => 'FOO', 'Naz' => 'BAZ', :default_foo => 'DEFAULT_FOO')
         end
         it "returns default JSON for version 4 that hasn't been declared" do
           example = JsonFoobar.create(foo: 'FOO', baz: 'BAZ', bar: 'BAR')
-          example.as_json(properties: :short, version: :v4).should == { foo: 'FOO', default_foo: 'DEFAULT_FOO' }
+          expect(example.as_json(properties: :short, version: :v4)).to eq(foo: 'FOO', default_foo: 'DEFAULT_FOO')
         end
         it 'returns JSON for the default version' do
           Mongoid::CachedJson.config.default_version = :v2
           example = JsonFoobar.create(foo: 'FOO', baz: 'BAZ', bar: 'BAR')
-          example.as_json(properties: :short).should == { :foo => 'FOO', 'Taz' => 'BAZ', 'Naz' => 'BAZ', :default_foo => 'DEFAULT_FOO' }
+          expect(example.as_json(properties: :short)).to eq(:foo => 'FOO', 'Taz' => 'BAZ', 'Naz' => 'BAZ', :default_foo => 'DEFAULT_FOO')
         end
         it 'returns correct JSON for Person used in README' do
           person = Person.create(first: 'John', middle: 'F.', last: 'Kennedy', born: 'May 29, 1917')
-          person.as_json.should == { name: 'John F. Kennedy' }
-          person.as_json(version: :v2).should == { first: 'John', middle: 'F.', last: 'Kennedy', name: 'John F. Kennedy' }
-          person.as_json(version: :v3).should == { first: 'John', middle: 'F.', last: 'Kennedy', name: 'John F. Kennedy', born: 'May 29, 1917' }
+          expect(person.as_json).to eq(name: 'John F. Kennedy')
+          expect(person.as_json(version: :v2)).to eq(first: 'John', middle: 'F.', last: 'Kennedy', name: 'John F. Kennedy')
+          expect(person.as_json(version: :v3)).to eq(first: 'John', middle: 'F.', last: 'Kennedy', name: 'John F. Kennedy', born: 'May 29, 1917')
         end
       end
       context 'polymorphic objects' do
@@ -404,18 +404,18 @@ describe Mongoid::CachedJson do
           end
         end
         it 'returns correct JSON when a child (embedded) polymorphic document is changed' do
-          @json_parent_foobar.as_json(properties: :all)[:json_polymorphic_embedded_foobar][:foo].should == 'embedded'
-          @json_embedded_foobar.as_json(properties: :all)[:foo].should == 'embedded'
+          expect(@json_parent_foobar.as_json(properties: :all)[:json_polymorphic_embedded_foobar][:foo]).to eq('embedded')
+          expect(@json_embedded_foobar.as_json(properties: :all)[:foo]).to eq('embedded')
           @json_embedded_foobar.update_attributes!(foo: 'EMBEDDED')
-          @json_embedded_foobar.as_json(properties: :all)[:foo].should == 'EMBEDDED'
-          @json_parent_foobar.as_json(properties: :all)[:json_polymorphic_embedded_foobar][:foo].should == 'EMBEDDED'
+          expect(@json_embedded_foobar.as_json(properties: :all)[:foo]).to eq('EMBEDDED')
+          expect(@json_parent_foobar.as_json(properties: :all)[:json_polymorphic_embedded_foobar][:foo]).to eq('EMBEDDED')
         end
         it 'returns correct JSON when a child (referenced) polymorphic document is changed' do
-          @json_parent_foobar.as_json(properties: :all)[:json_polymorphic_referenced_foobar][:foo].should == 'referenced'
-          @json_referenced_foobar.as_json(properties: :all)[:foo].should == 'referenced'
+          expect(@json_parent_foobar.as_json(properties: :all)[:json_polymorphic_referenced_foobar][:foo]).to eq('referenced')
+          expect(@json_referenced_foobar.as_json(properties: :all)[:foo]).to eq('referenced')
           @json_referenced_foobar.update_attributes!(foo: 'REFERENCED')
-          @json_referenced_foobar.as_json(properties: :all)[:foo].should == 'REFERENCED'
-          @json_parent_foobar.as_json(properties: :all)[:json_polymorphic_referenced_foobar][:foo].should == 'REFERENCED'
+          expect(@json_referenced_foobar.as_json(properties: :all)[:foo]).to eq('REFERENCED')
+          expect(@json_parent_foobar.as_json(properties: :all)[:json_polymorphic_referenced_foobar][:foo]).to eq('REFERENCED')
         end
       end
       context 'polymorhphic relationships' do
@@ -426,14 +426,14 @@ describe Mongoid::CachedJson do
           @person_post = PolyPost.create!(postable: @person)
         end
         it 'returns the correct JSON' do
-          @company_post.as_json.should == { parent: { id: @company.id, type: 'PolyCompany' } }
-          @person_post.as_json.should == { parent: { id: @person.id, type: 'PolyPerson' } }
+          expect(@company_post.as_json).to eq(parent: { id: @company.id, type: 'PolyCompany' })
+          expect(@person_post.as_json).to eq(parent: { id: @person.id, type: 'PolyPerson' })
         end
       end
       context 'cache key' do
         it 'correctly generates a cached json key' do
           example = JsonFoobar.create(foo: 'FOO', baz: 'BAZ', bar: 'BAR')
-          JsonFoobar.cached_json_key({ properties: :short, is_top_level_json: true, version: :v1 }, example.class, example.id).should == "as_json/v1/JsonFoobar/#{example.id}/short/true"
+          expect(JsonFoobar.cached_json_key({ properties: :short, is_top_level_json: true, version: :v1 }, example.class, example.id)).to eq("as_json/v1/JsonFoobar/#{example.id}/short/true")
         end
       end
       context 'embeds_many relationships' do
@@ -443,12 +443,12 @@ describe Mongoid::CachedJson do
           @cell.inmates.create!(nickname: 'Bob', person: Person.create!(first: 'Bob'))
         end
         it 'returns the correct JSON' do
-          @cell.as_json(properties: :all).should == { number: 42,
-                                                      inmates: [
-                                                        { nickname: 'Joe', person: { name: 'Joe' } },
-                                                        { nickname: 'Bob', person: { name: 'Bob' } }
-                                                      ]
-          }
+          expect(@cell.as_json(properties: :all)).to eq(number: 42,
+                                                        inmates: [
+                                                          { nickname: 'Joe', person: { name: 'Joe' } },
+                                                          { nickname: 'Bob', person: { name: 'Bob' } }
+                                                        ]
+          )
         end
       end
       context 'with repeated objects in the JSON' do
@@ -459,12 +459,12 @@ describe Mongoid::CachedJson do
           @cell.inmates.create!(nickname: 'Bob', person: @person)
         end
         it 'returns the correct JSON' do
-          @cell.as_json(properties: :all).should == { number: 42,
-                                                      inmates: [
-                                                        { nickname: 'Joe', person: { name: 'Evil' } },
-                                                        { nickname: 'Bob', person: { name: 'Evil' } }
-                                                      ]
-          }
+          expect(@cell.as_json(properties: :all)).to eq(number: 42,
+                                                        inmates: [
+                                                          { nickname: 'Joe', person: { name: 'Evil' } },
+                                                          { nickname: 'Bob', person: { name: 'Evil' } }
+                                                        ]
+          )
         end
       end
       context 'belongs_to relationship' do
@@ -472,7 +472,7 @@ describe Mongoid::CachedJson do
           @tool = Tool.create!(name: 'hammer')
         end
         it 'returns a nil reference' do
-          @tool.as_json(properties: :all).should == { tool_box: nil, name: 'hammer' }
+          expect(@tool.as_json(properties: :all)).to eq(tool_box: nil, name: 'hammer')
         end
         context 'persisted' do
           before :each do
@@ -480,7 +480,7 @@ describe Mongoid::CachedJson do
             @tool.update_attributes!(tool_box: @tool_box)
           end
           it 'returns a reference' do
-            @tool.as_json(properties: :all).should == { tool_box: { color: 'red' }, name: 'hammer' }
+            expect(@tool.as_json(properties: :all)).to eq(tool_box: { color: 'red' }, name: 'hammer')
           end
         end
       end
@@ -489,11 +489,11 @@ describe Mongoid::CachedJson do
           @image = FastJsonImage.create!
         end
         it 'resolves a default empty relationship' do
-          @image.as_json(properties: :all).should == { name: 'Image', urls: [] }
+          expect(@image.as_json(properties: :all)).to eq(name: 'Image', urls: [])
         end
         it 'resolves a nil relationship on destroy' do
           @image.destroy
-          @image.as_json(properties: :all).should == { name: 'Image', urls: [] }
+          expect(@image.as_json(properties: :all)).to eq(name: 'Image', urls: [])
         end
       end
     end