+++ spec/dalli_spec.rb
@@ -12,7 +12,7 @@ describe ActiveSupport::Cache::DalliStore do
     Mongoid::CachedJson::Config.cache = @cache
   end
   it 'uses dalli_store' do
-    Mongoid::CachedJson.config.cache.should be_a ActiveSupport::Cache::DalliStore
+    expect(Mongoid::CachedJson.config.cache).to be_a ActiveSupport::Cache::DalliStore
   end
   context 'read_multi' do
     context 'array' do
@@ -22,15 +22,15 @@ describe ActiveSupport::Cache::DalliStore do
         tool1_key = Tool.cached_json_key(options, Tool, tool1.id)
         tool2 = Tool.create!(name: 'screwdriver')
         tool2_key = Tool.cached_json_key(options, Tool, tool2.id)
-        Mongoid::CachedJson.config.cache.should_not_receive(:fetch)
-        Mongoid::CachedJson.config.cache.should_receive(:read_multi).with(tool1_key, tool2_key).once.and_return(
+        expect(Mongoid::CachedJson.config.cache).not_to receive(:fetch)
+        expect(Mongoid::CachedJson.config.cache).to receive(:read_multi).with(tool1_key, tool2_key).once.and_return(
           tool1_key => { _id: tool1.id.to_s },
           tool2_key => { _id: tool2.id.to_s }
         )
-        [tool1, tool2].as_json(properties: :all).should == [
+        expect([tool1, tool2].as_json(properties: :all)).to eq([
           { tool_box: nil, _id: tool1.id.to_s },
           { tool_box: nil, _id: tool2.id.to_s }
-        ]
+        ])
       end
     end
   end